Restructured backend classes
