+++ backend/src/main/java/pl/lodz/p/it/opinioncollector/exceptions/PasswordNotMatchesException.java
@@ -1,4 +1,4 @@
-package pl.lodz.p.it.opinioncollector.userModule.user;
+package pl.lodz.p.it.opinioncollector.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
