+++ backend/src/main/java/pl/lodz/p/it/opinioncollector/userModule/dto/RegisterUserDTO.java
@@ -1,4 +1,4 @@
-package pl.lodz.p.it.opinioncollector.userModule.auth;
+package pl.lodz.p.it.opinioncollector.userModule.dto;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
