+++ backend/src/main/java/pl/lodz/p/it/opinioncollector/userModule/user/UserManager.java
@@ -9,6 +9,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
+import pl.lodz.p.it.opinioncollector.exceptions.PasswordNotMatchesException;
 import pl.lodz.p.it.opinioncollector.userModule.auth.*;
 import pl.lodz.p.it.opinioncollector.userModule.token.Token;
 import pl.lodz.p.it.opinioncollector.userModule.token.TokenRepository;
@@ -59,7 +60,7 @@ public class UserManager implements UserDetailsService {
             String link = "http://localhost:4200/resetConfirm/" + resetPasswordToken;
             mailManager.resetPasswordEmail(email, "your email " + email, link);
         } catch (Exception e) {
-            throw new EmailAlreadyRegisteredException();
+            throw new ResponseStatusException(HttpStatus.CONFLICT);
         }
     }
 
@@ -74,7 +75,6 @@ public class UserManager implements UserDetailsService {
 
     public void changePassword(String oldPassword, String newPassword) throws PasswordNotMatchesException {
         //TODO check if password is strong enough
-
         String email = SecurityContextHolder.getContext().getAuthentication().getName();
         User user = loadUserByUsername(email);
         if (encoder.matches(oldPassword, user.getPassword())) {
@@ -94,7 +94,7 @@ public class UserManager implements UserDetailsService {
             tokenRepository.deleteTokenByUser(user);
             mailManager.adminActionEmail(user.getEmail(), user.getUsername(), "blocked");
         } catch (Exception e) {
-            throw new EmailAlreadyRegisteredException();
+            throw new ResponseStatusException(HttpStatus.CONFLICT);
         }
     }
 
@@ -106,7 +106,7 @@ public class UserManager implements UserDetailsService {
             userRepository.save(user);
             mailManager.adminActionEmail(user.getEmail(), user.getVisibleName(), "unlocked");
         } catch (Exception e) {
-            throw new EmailAlreadyRegisteredException();
+            throw new ResponseStatusException(HttpStatus.CONFLICT);
         }
     }
 
@@ -121,7 +121,7 @@ public class UserManager implements UserDetailsService {
             String link = "http://localhost:8080/api/confirm/remove?token=" + deletionToken;
             mailManager.deletionEmail(user.getEmail(), user.getVisibleName(), link);
         } catch (Exception e) {
-            throw new EmailAlreadyRegisteredException();
+            throw new ResponseStatusException(HttpStatus.CONFLICT);
         }
     }
 
@@ -133,15 +133,19 @@ public class UserManager implements UserDetailsService {
             tokenRepository.deleteTokenByUser(user);
             mailManager.adminActionEmail(user.getEmail(), user.getVisibleName(), "deleted");
         } catch (Exception e) {
-            throw new EmailAlreadyRegisteredException();
+            throw new ResponseStatusException(HttpStatus.CONFLICT);
         }
     }
 
     public void changeUsername(String newUsername) {
-        String email = SecurityContextHolder.getContext().getAuthentication().getName();
-        User user = loadUserByUsername(email);
-        user.setVisibleName(newUsername);
-        userRepository.save(user);
+        try {
+            String email = SecurityContextHolder.getContext().getAuthentication().getName();
+            User user = loadUserByUsername(email);
+            user.setVisibleName(newUsername);
+            userRepository.save(user);
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT);
+        }
     }
 
     public List<User> getAllUsers(String email) {;
